% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{check_within_Window}
\alias{check_within_Window}
\title{Internal function to check if a set of points is
within a given window}
\usage{
check_within_Window(x, y, w)
}
\arguments{
\item{x}{numeric vector, x coordinates of the points}

\item{y}{numeric vector, y coordinates of the points}

\item{w}{owin class object, observation window}
}
\value{
numeric vector of ones and zeros with
one meaning the point is within the window
}
\description{
Internal function to check if a set of points is
within a given window
}
