% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getbignum2}
\alias{getbignum2}
\title{Internal function to compute the sum of the log terms}
\usage{
getbignum2(Y_x, Y_y, C_x, C_y, h)
}
\arguments{
\item{Y_x}{numeric vector, x coordinates of the offspring process}

\item{Y_y}{numeric vector, y coordinates of the offspring process}

\item{C_x}{numeric vector, x coordinates of the parent process}

\item{C_y}{numeric vector, y coordinates of the parent process}

\item{h}{positive scalar, bandwidth}
}
\value{
real scalar
}
\description{
Internal function to compute the sum of the log terms
}
