% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{gethpars}
\alias{gethpars}
\title{Internal function to get initial values for the bandwidth parameters}
\usage{
gethpars(Y_x, Y_y, C_x, C_y)
}
\arguments{
\item{Y_x}{numeric vector, x coordinates of the offspring process}

\item{Y_y}{numeric vector, y coordinates of the offspring process}

\item{C_x}{numeric vector, x coordinates of the parent process}

\item{C_y}{numeric vector, y coordinates of the parent process}
}
\value{
positive scalar for initial value of h
}
\description{
Internal function to get initial values for the bandwidth parameters
}
