% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{MCPP_run}
\alias{MCPP_run}
\title{R function to draw posterior samples from MCPP}
\usage{
MCPP_run(
  obj,
  parent_genus,
  offspring_genus,
  bdtype = 1,
  B = 100,
  jitter = FALSE,
  hyperParams = list(lambda = c(0.01, 0.01), alpha = c(0.01, 0.01), lambdaO = c(0.01,
    0.01), h = 0.05),
  startValues = NULL,
  mcmcParams = list(run = list(numReps = 20000, thin = 1, burninPerc = 0.5), storage =
    NULL, tuning = NULL),
  store_res = FALSE,
  path = NULL
)
}
\arguments{
\item{obj}{ppp class object, the dataset}

\item{parent_genus}{character vector, scalar if all parent types are same or a vector of name of the parent genuses}

\item{offspring_genus}{character vector, name of the offspring genuses}

\item{bdtype}{numeric scalar, can take values 1 or 2. 1 keeps the window as is; 2 creates a convex hull and uses that as the window}

\item{B}{positive integer, number of Monte Carlo samples to use to approximate the integral term}

\item{jitter}{logical, if TRUE, the initial value for the bandwidth parameter is jittered. Useful for running multiple chains}

\item{hyperParams}{named list of hyperparameters needed for the MCPP model, components are, lambda (two-vector of shape and rate hyperparamaters for the Gamma prior to the parent intensities), alpha (two-vector of shape and rate hyperparamaters for the Gamma prior to the alpha parameter of the offspring densities), lambdaO (two-vector of shape and rate hyperparamaters for the Gamma prior to the unrelated taxa intensities), h (scalar proportion of the maximum window distance to be set equal to the 99th percentile of the Half-normal prior to the bandwidth parameter)}

\item{startValues}{named list of starting values for the alpha and h parameters, components are, alpha (positive vector of initial values for the alpha parameter, same size as the number of offspring taxa) and h (positive vector of initial values for the bandwidth parameter, same size as the number of offspring taxa). Default is NULL in which case initial values are computed}

\item{mcmcParams}{named list of parameters required for MCMC sampling, components are, run (named list with entries numReps: total number of scans, thin: the thinning interval length and burninPerc: percentage of scans to be considered as burn-in), storage (ignored for MCPP) and tuning (ignored for MCPP)}

\item{store_res}{logical, indicator to store results periodically}

\item{path}{path to the storage file where the periodical results are to be saved}
}
\value{
list of posterior samples for the corresponding parameters
}
\description{
R function to draw posterior samples from MCPP
}
