% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{NS_MCMC_new_fixp}
\alias{NS_MCMC_new_fixp}
\title{R function to draw posterior samples from MACPP}
\usage{
NS_MCMC_new_fixp(
  obj,
  parent_genus,
  offspring_genus,
  bdtype = 1,
  jitter = FALSE,
  al = 0.01,
  bl = 0.01,
  am = 0.01,
  bm = 0.01,
  ao = 0.01,
  bo = 0.01,
  hclimp = 0.05,
  B = 100,
  iters = 20000,
  burn = 10000,
  thin = 1,
  step_int = 30,
  store_res = FALSE,
  outfile = NULL
)
}
\arguments{
\item{obj}{ppp class object, the dataset}

\item{parent_genus}{character vector, scalar if all parent types are same or a vector of name of the parent genuses}

\item{offspring_genus}{character vector, name of the offspring genuses}

\item{bdtype}{numeric scalar, can take values 1 or 2. 1 keeps the window as is; 2 creates a convex hull and uses that as the window}

\item{jitter}{logical, if TRUE, the initial value for the bandwidth parameter is jittered. Useful for running multiple chains}

\item{al}{positive scalar, shape hyperparameter for the Gamma priors to the parent intensities}

\item{bl}{positive scalar, rate hyperparameter for the Gamma priors to the parent intensities}

\item{am}{positive scalar, shape hyperparameter for the Gamma priors to the offspring densities}

\item{bm}{positive scalar, rate hyperparameter for the Gamma priors to the offspring densities}

\item{ao}{positive scalar, shape hyperparameter for the Gamma priors to the unrelated taxon intensities}

\item{bo}{positive scalar, rate hyperparameter for the Gamma priors to the unrelated taxon intensities}

\item{hclimp}{proportion between 0 and 1, proportion of the maximum window distance to be set equal to the 99th percentile of the Half-normal prior to the bandwidth parameter}

\item{B}{positive integer, number of Monte Carlo samples to use to approximate the integral term}

\item{iters}{positive integer, number of iterations of the entire MCMC}

\item{burn}{positive integer, number of samples to be treated as burn-in period samples and discarded}

\item{thin}{positive integer, the thinning frequency}

\item{step_int}{positive integer, unused}

\item{store_res}{logical, indicator to store results periodically}

\item{outfile}{path, path to the storage file where the periodical results are to be saved}
}
\value{
list of posterior samples for the corresponding parameters
}
\description{
R function to draw posterior samples from MACPP
}
