% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.MCPP}
\alias{print.MCPP}
\title{Print function for MCPP class objects}
\usage{
\method{print}{MCPP}(x, digits = 3, ci = 0.95, ...)
}
\arguments{
\item{x}{object of class MCPP}

\item{digits}{number of digits to print}

\item{ci}{level of confidence interval, between 0 and 1}

\item{...}{additional parameters supplied to the function}
}
\value{
prints outputs from the object
}
\description{
Print function for MCPP class objects
}
