% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summ.MCPP}
\alias{print.summ.MCPP}
\title{Print function for summary of MCPP class objects}
\usage{
\method{print}{summ.MCPP}(x, digits = 3, ...)
}
\arguments{
\item{x}{summary object of class MCPP}

\item{digits}{number of digits to print}

\item{...}{additional parameters supplied to the function}
}
\value{
prints outputs from the summary object
}
\description{
Print function for summary of MCPP class objects
}
