% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{subdata}
\alias{subdata}
\title{Internal function to subset the data based on genus name}
\usage{
subdata(dat, genus)
}
\arguments{
\item{dat}{ppp class object, the entire dataset}

\item{genus}{character, name of the genus to be subsetted}
}
\value{
ppp class object, subset of the data with only the specified genus
}
\description{
Internal function to subset the data based on genus name
}
