% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.MCPP}
\alias{summary.MCPP}
\title{Summary function for MCPP class objects}
\usage{
summary.MCPP(x, digits = 3, ci = 0.95, ...)
}
\arguments{
\item{x}{object of class MCPP}

\item{digits}{number of digits to print}

\item{ci}{level of confidence interval, between 0 and 1}

\item{...}{additional parameters supplied to the function}
}
\value{
Summary outputs from the object
}
\description{
Summary function for MCPP class objects
}
