% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiprocppp.R
\name{multiprocppp}
\alias{multiprocppp}
\title{R function to generate from multiprocess poisson point process}
\usage{
multiprocppp(
  win,
  lambdapar,
  name_par = NULL,
  noff,
  mu0vec,
  hvec,
  kern = "Gaussian",
  names_off = NULL,
  nxtra = 0,
  lambdaxtra,
  names_xtra = NULL
)
}
\arguments{
\item{win}{owin class object, the observation window}

\item{lambdapar}{positive scalar, the intensity parameter for the parent process}

\item{name_par}{optional character, name of the parent}

\item{noff}{positive scalar, number of offsprings}

\item{mu0vec}{numeric vector of positive numbers, vector of offspring densities}

\item{hvec}{numeric vector of positive numbers, vector of bandwidth parameters}

\item{kern}{character, choice of kernel function. Permitted values are "Gaussian", "Uniform" or "Cauchy"}

\item{names_off}{character vector, names of the offspring entities}

\item{nxtra}{non-negative integer, number of unrelated entities}

\item{lambdaxtra}{scalar or vector of positive numbers, intensities of the unrelated entities processes}

\item{names_xtra}{scalar or vector of characters, names of the unrelated entities}
}
\value{
object of class ppp
}
\description{
R function to generate from multiprocess poisson point process
}
